create doctor model/service and change admin service
