+++ src/config/database.ts
@@ -8,6 +8,7 @@ import { TenantExams } from '../models/TenantExams';
 import { PatientExams } from '../models/PatientExams';
 import {NoticeCard} from "../models/NoticeCard";
 import {PatientClinic} from "../models/PatientClinic";
+import {Doctor} from "../models/Doctor";
 
 dotenv.config();
 
@@ -18,7 +19,7 @@ export const AppDataSource = new DataSource({
     username: process.env.DB_USERNAME,
     password: process.env.DB_PASSWORD,
     database: process.env.DB_NAME,
-    entities: [Admin, NoticeCard, Patient, PatientClinic, Tenant, Product, TenantExams, PatientExams],
+    entities: [Admin, NoticeCard, Patient, PatientClinic, Tenant, Product, TenantExams, PatientExams, Doctor],
     synchronize: true,
     logging: true,
     ssl: process.env.NODE_ENV === 'production' ? { rejectUnauthorized: false } : false,
