+++ src/controllers/adminController.ts
@@ -1,17 +1,13 @@
 import {Request, Response} from 'express';
 import {customErrorResponse, errorResponse, successResponse} from '../utils/httpResponses';
-import {getOnlyAdmins, getAdminByCPF, getAdminsByName, getDoctors, getDoctorsByExamName, updateAdmin, deleteAdmin} from "../services/adminService";
+import {
+    getAdminByCPF,
+    getAdminsByName,
+    updateAdmin,
+    deleteAdmin,
+    getAdmins
+} from "../services/adminService";
 
-interface PaginationQuery {
-    page?: string;
-    take?: string;
-    skip?: string;
-}
-
-interface GetDoctorsResult {
-    doctors: any[];
-    total: number;
-}
 
 export const getAdminListController =  async (req: Request, res: Response) => {
     /*
@@ -21,63 +17,14 @@ export const getAdminListController =  async (req: Request, res: Response) => {
     */
     try {
         const tenantId = req.headers['x-tenant-id'];
-        const result = await getOnlyAdmins(parseInt(tenantId as string));
+        const result = await getAdmins(parseInt(tenantId as string));
         return successResponse(res, result);
 
     } catch (error) {
         return errorResponse(res, error);
     }
 }
-export const getDoctorsListController = async (req: Request, res: Response) => {
-    /*
-    #swagger.tags = ['Admin']
-    #swagger.summary = 'List All Doctors by Tenant with pagination'
-    #swagger.description = 'Get All Doctors from a Tenant with pagination By default list 10'
-    */
-    try {
-        const tenantId = req.headers['x-tenant-id'];
-        if (!tenantId || typeof tenantId !== 'string') {
-            return errorResponse(res, new Error('Tenant ID inválido ou não informado'), 400);
-        }
 
-        const { page = '1', take = '10', skip = '0' } = req.query as PaginationQuery;
-
-        const numericParams = {
-            tenantId: parseInt(tenantId),
-            take: parseInt(take),
-            skip: parseInt(skip),
-            page: parseInt(page)
-        };
-
-        if (Object.values(numericParams).some(isNaN)) {
-            return errorResponse(res, new Error('Invalid pagination parameters'), 400);
-        }
-        const result: GetDoctorsResult = await getDoctors({
-            tenantId: numericParams.tenantId,
-            take: numericParams.take,
-            skip: numericParams.skip
-        });
-
-        const remaining = result.total - result.doctors.length;
-
-        const message = `Mostrando ${result.doctors.length} de ${result.total} médicos (${remaining} faltando)`;
-
-        return successResponse(res, {
-            data: result.doctors,
-            pagination: {
-                total: result.total,
-                page: numericParams.page,
-                take: numericParams.take,
-                skip: numericParams.skip,
-                remaining
-            }
-        }, message);
-
-    } catch (error) {
-        console.error('Error in getDoctorsListController', error);
-        return errorResponse(res, error);
-    }
-};
 
 export const getAdminsByCPFController = async (req: Request, res: Response) => {
     /*
@@ -117,24 +64,8 @@ export const getAdminsByNameController = async (req: Request, res: Response) => 
    }
 }
 
-export const getDoctorsByExamNameController = async (req: Request, res: Response) => {
-    /*
-    #swagger.tags = ['Admin']
-    #swagger.summary = 'Get Doctors by Exam  '
-    #swagger.description = 'Filter Doctors by exam name'
-    */
-    try {
-        const { examName } = req.query;
-        const result = await getDoctorsByExamName(examName as string);
-        return successResponse(res, result, 'Doutores associados ao exame listados com sucesso');
-    } catch (error) {
-        return errorResponse(res, error);
-    }
-};
 
-/**
- * Atualiza informações de um admin existente
- */
+
 export const updateAdminController = async (req: Request, res: Response) => {
     /*
     #swagger.tags = ['Admin']
@@ -143,9 +74,9 @@ export const updateAdminController = async (req: Request, res: Response) => {
     */
     try {
         const adminId = parseInt(req.params.id);
-        const { email, fullName, cpf, CRM, phone, occupation } = req.body;
+        const { email, fullName, cpf, cep, phone } = req.body;
         
-        const result = await updateAdmin(adminId, { email, fullName, cpf, CRM, phone, occupation });
+        const result = await updateAdmin(adminId, { email, fullName, cpf, phone, cep });
         return successResponse(res, result, 'Admin atualizado com sucesso');
     } catch (error) {
         return errorResponse(res, error);
