+++ src/repositories/doctorRepository.ts
@@ -1,0 +1,4 @@
+import { AppDataSource } from "../config/database";
+import {Doctor} from "../models/Doctor";
+
+export const doctorRepository = AppDataSource.getRepository(Doctor);
