+++ src/types/dto/auth/registerAdminDTO.ts
@@ -2,9 +2,8 @@ export interface RegisterAdminDTO {
     email: string;
     cpf: string;
     fullName: string;
-    CRM?: string;
     phone?: string;
-    isDoctor?: boolean;
+    role?: string;
+    cep?: string;
     password?: string;
-    occupation?: string
 }
